+++ mongo/request_manager.go
@@ -595,15 +595,6 @@ func (r *RequestManager) RevokeAccessToken(ctx context.Context, requestID string
 		return err
 	}
 
-	err = r.Cache.Delete(ctx, storage.EntityCacheAccessTokens, cacheObject.Key())
-	if err != nil {
-		// Log to StdOut
-		log.WithError(err).Error(logError)
-		// Log to OpenTracing
-		otLogErr(span, err)
-		return err
-	}
-
 	return nil
 }